GetImplementationTasksTest passes
